Add pluggable docker/buildah container image engines
Extract the container builder's toolchain into a shared ContainerEngine
interface with docker (local dev) and buildah (Vercel build container)
implementations. Select buildah when VERCEL_BUILD_IMAGE is set; otherwise
use docker. Shared VCR auth, repository creation, and readiness logic stay
in the orchestrator.
